LayoutTests/media/wireless-playback-media-player/remote-playback-connecting-expected.txt


Test that a 'connecting' event is dispatched when a MediaDeviceRoute activates.

EXPECTED (video.remote.state == 'disconnected') OK
RUN(video.src = findMediaFile('video', '../content/test'))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
RUN(internals.mockMediaDeviceRouteController.activateRoute(route))
EVENT(connecting)
END OF TEST
